requirements.txt
torch
huggingface_hub
transformers
sentencepiece
mysql-connector-python
flask
flask-cors
pandas
openpyxl
xlwt
